Update README.md
I am forced to delete my statement because a new law has been introduced:

"Article 207.3 "Public dissemination of knowingly false information about the use of the Armed Forces of the Russian Federation" appeared in the Criminal Code of the Russian Federation."

Which speaks of the punishment for calling - to stop the war.

" For appeals to stop the war - first an administrative article, then a criminal article. If the call caused damage, the criminal article at once."

Since we can't count on a fair trial in Russia, I'm retracting my post. 
(I'm afraid of being prosecuted for my opinion)

But I am still against the war!